+++ lib/graphql/language/printer.rb
@@ -194,10 +194,8 @@ module GraphQL
         print_string("$#{variable_identifier.name}")
       end
 
-
-      def print_schema_definition(schema)
-        has_conventional_names =
-          (schema.query.nil? || schema.query == 'Query') &&
+      def print_schema_definition(schema, extension: false)
+        has_conventional_names = (schema.query.nil? || schema.query == 'Query') &&
           (schema.mutation.nil? || schema.mutation == 'Mutation') &&
           (schema.subscription.nil? || schema.subscription == 'Subscription')
 
@@ -205,7 +203,7 @@ module GraphQL
           return
         end
 
-        print_string("schema")
+        extension ? print_string("extend schema") : print_string("schema")
 
         if schema.directives.any?
           schema.directives.each do |dir|
@@ -230,14 +228,15 @@ module GraphQL
         end
       end
 
-      def print_scalar_type_definition(scalar_type)
-        print_description(scalar_type)
+
+      def print_scalar_type_definition(scalar_type, extension: false)
+        extension ? print_string("extend ") : print_description(scalar_type)
         print_string("scalar #{scalar_type.name}")
         print_directives(scalar_type.directives)
       end
 
-      def print_object_type_definition(object_type)
-        print_description(object_type)
+      def print_object_type_definition(object_type, extension: false)
+        extension ? print_string("extend ") : print_description(object_type)
         print_string("type #{object_type.name}")
         print_implements(object_type) unless object_type.interfaces.empty?
         print_directives(object_type.directives)
@@ -289,23 +288,23 @@ module GraphQL
         print_directives(field.directives)
       end
 
-      def print_interface_type_definition(interface_type)
-        print_description(interface_type)
+      def print_interface_type_definition(interface_type, extension: false)
+        extension ? print_string("extend ") : print_description(interface_type)
         print_string("interface #{interface_type.name}")
         print_implements(interface_type) if interface_type.interfaces.any?
         print_directives(interface_type.directives)
         print_field_definitions(interface_type.fields)
       end
 
-      def print_union_type_definition(union_type)
-        print_description(union_type)
+      def print_union_type_definition(union_type, extension: false)
+        extension ? print_string("extend ") : print_description(union_type)
         print_string("union #{union_type.name}")
         print_directives(union_type.directives)
         print_string(" = #{union_type.types.map(&:name).join(" | ")}")
       end
 
-      def print_enum_type_definition(enum_type)
-        print_description(enum_type)
+      def print_enum_type_definition(enum_type, extension: false)
+        extension ? print_string("extend ") : print_description(enum_type)
         print_string("enum #{enum_type.name}")
         print_directives(enum_type.directives)
         print_string(" {\n")
@@ -322,8 +321,8 @@ module GraphQL
         print_string("\n")
       end
 
-      def print_input_object_type_definition(input_object_type)
-        print_description(input_object_type)
+      def print_input_object_type_definition(input_object_type, extension: false)
+        extension ? print_string("extend ") : print_description(input_object_type)
         print_string("input #{input_object_type.name}")
         print_directives(input_object_type.directives)
         if !input_object_type.fields.empty?
@@ -429,24 +428,38 @@ module GraphQL
           print_variable_identifier(node)
         when Nodes::SchemaDefinition
           print_schema_definition(node)
+        when Nodes::SchemaExtension
+          print_schema_definition(node, extension: true)
         when Nodes::ScalarTypeDefinition
           print_scalar_type_definition(node)
+        when Nodes::ScalarTypeExtension
+          print_scalar_type_definition(node, extension: true)
         when Nodes::ObjectTypeDefinition
           print_object_type_definition(node)
+        when Nodes::ObjectTypeExtension
+          print_object_type_definition(node, extension: true)
         when Nodes::InputValueDefinition
           print_input_value_definition(node)
         when Nodes::FieldDefinition
           print_field_definition(node)
         when Nodes::InterfaceTypeDefinition
           print_interface_type_definition(node)
+        when Nodes::InterfaceTypeExtension
+          print_interface_type_definition(node, extension: true)
         when Nodes::UnionTypeDefinition
           print_union_type_definition(node)
+        when Nodes::UnionTypeExtension
+          print_union_type_definition(node, extension: true)
         when Nodes::EnumTypeDefinition
           print_enum_type_definition(node)
+        when Nodes::EnumTypeExtension
+          print_enum_type_definition(node, extension: true)
         when Nodes::EnumValueDefinition
           print_enum_value_definition(node)
         when Nodes::InputObjectTypeDefinition
           print_input_object_type_definition(node)
+        when Nodes::InputObjectTypeExtension
+          print_input_object_type_definition(node, extension: true)
         when Nodes::DirectiveDefinition
           print_directive_definition(node)
         when FalseClass, Float, Integer, NilClass, String, TrueClass, Symbol